Add files via upload
Code and analysis of how past years cone counts (in trees) influence parturition dates. The idea was to see how far into the past the cone crop matters for current resource dependent traits. The previous year (t-1) is the most recent relevant measure of resource availability, then the year before that (t-2) and so on. 
Results indicate t-1 and t-2 are equally important, while t-3 has about 1/3 of the influence of those years.

Possibly suggests an index of current resource availability should be cones in t-1 + cones in t-2 + 1/3cones t-3.