Add files via upload
Code to select and plot territorial behaviours by individual squirrel on zoomed-in area of grid. Intended to be used in the field by connected to the local instance of the database at camp. The resulting maps are intended to be brought into the field when mapping squirrel territory boundaries with GPS tracks following 2017 protocol.